// web/src/api/generated/schemas/courseResource.zod.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Fonzytooter API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from 'zod'
import { ModuleSummary } from './moduleSummary.zod'

export const CourseResource = zod.strictObject({
  $schema: zod.url().optional().describe('A URL to the JSON Schema for this object.'),
  description: zod.string(),
  id: zod.string(),
  modules: zod.array(ModuleSummary),
  order: zod.int(),
  title: zod.string(),
})

export type CourseResource = zod.input<typeof CourseResource>
export type CourseResourceOutput = zod.output<typeof CourseResource>